+++ test/tests.cpp
@@ -1,12 +1,11 @@
-// Copyright 2020 GHA Test Team
+// Copyright 2020 UNN-IASR
 #include "fun.h"
 
 #include <gtest/gtest.h>
 
 TEST(MyFunTest, test1) {
-  long long x=2;
-  unsigned short n=8;
-  long long result=power(x,n);
-  EXPECT_EQ(256,result);
+  int64_t x = 2;
+  uint16_t n = 8;
+  uint64_t result = power(x, n);
+  EXPECT_EQ(256, result);
 }
-
